requirements.txt
alembic==1.6.5
amqp==5.0.6
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
attrs==17.4.0
auger-python==0.1.35
autopep8==1.5.7
beautifulsoup4==4.6.0
behave==1.2.6
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.2
certifi==2018.1.18
chardet==3.0.4
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==2.1.4
cryptography-vectors==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==5.0.9
defer==1.0.6
dnspython==2.1.0
email-validator==1.1.2
face-recognition==1.4.0
face-recognition-models==0.3.0
feedparser==5.2.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
flask-sock==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
funcsigs==1.0.2
greenlet==1.1.0
gunicorn==20.0.4
h11==0.12.0
html5lib==0.999999999
httplib2==0.9.2
idna==3.2
importlib-metadata==4.6.4
imutils==0.5.4
itsdangerous==2.0.1
jetson-stats==3.1.0
Jetson.GPIO==2.0.16
Jinja2==3.0.1
joblib==0.11
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lxml==4.2.1
Mako==1.1.4
MarkupSafe==2.0.1
mock==4.0.3
nose==1.3.7
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pandas==0.22.0
parse==1.19.0
parse-type==0.5.2
pathlib==1.0.1
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==3.0.19
protobuf==3.0.0
psutil==5.8.0
psycopg2==2.7.4
py==1.5.2
pycairo==1.16.2
pycodestyle==2.7.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pytest==3.3.2
pytz==2021.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==22.0.3
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.19.1
scipy==0.19.1
SecretStorage==2.3.1
selenium==3.141.0
semantic-version==2.8.5
setuptools-rust==0.12.1
simple-websocket==0.3.0
simplejson==3.13.2
six==1.16.0
somewhere==0.0.1
SQLAlchemy==1.4.14
ssh-import-id==5.7
super-state-machine==2.0.2
system-service==0.3
systemd-python==234
tensorrt==7.1.3.0
toml==0.10.2
Turbo-Flask==0.6.0
typing-extensions==3.10.0.0
uff==0.6.9
urllib3==1.26.6
urwid==2.0.1
uuid==1.30
validators==0.18.2
vine==5.0.0
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5
Werkzeug==2.0.1
wget==3.2
wsproto==1.0.0
WTForms==2.3.3
xkit==0.0.0
youtube-dl==2018.3.14
zipp==3.5.0
zope.interface==4.3.2
